New package: app-crypt/keybase-bin-2.13.0
Since building Electron apps in Portage is decidedly nontrivial, pull
in upstream's Debian package as a -bin ebuild to get the GUI.

GUI and CLI works. KBFS works, but mounts in the wrong place
(/var/run/user/$UID/keybase/kbfs/). Firefox and Chrome/Chromium
integration is not tested, and is currently not installed.

Only tested on amd64, but should work on x86.